migrate to tsx, add publish scripts, add package metadata
- replace ts-node/nodemon with tsx
- extract vsceRun helper, refactor package-vsix and add publish:vsce/ovsx scripts
- remove GitHub Actions publish workflow in favour of manual publish scripts
- add author and homepage fields
