+++ tsconfig.json
@@ -11,11 +11,6 @@
     "outDir": "./dist",
     "rootDir": "./src"
   },
-  "ts-node": {
-    "compilerOptions": {
-      "module": "CommonJS"
-    }
-  },
   "include": ["src/**/*"],
   "exclude": ["node_modules"]
 }
